Compatibility with Make for Windows
From testing on a Windows 8.1 machine with MinGW installed, it seems as
though cmd.exe can handle paths with either back/front slashes, but some
cases break:
- `xargs` strips the backslashes from the command.
- cannot use current directory (`./`), produces error:
`'.' is not recognized as an internal or external command, operable
program or batch file.`

Solution used here is to continue to use slashes, but convert all
backslashes to slashes where possible. Also, convert current directory
wherever used to `.\` in Windows. We could use a {$CURR_DIR} instead,
but wanted to keep it more streamlined for the non Windows devs, which
is the large majority.

This commit should not change any behaviour on Linux and Mac, but was
not tested on them.