Give better feedback on bad URLs in settings dialog.
This is a fix for https://github.com/medic/medic-webapp/issues/1077

When the `save` button is clicked on the settings dialog, the URL is contacted
to check if it is a valid couch database.  If it's a couch server, but not a db,
we try looking for a db named `medic` on that server.  If we can't find a valid
db, the user is shown an error relating to the problem.