Replace the standard WebView with a fork of crosswalk to allow unlimited storage
The standard WebView uses a device-specific version of Chrome/Chromium,
depending mainly on the version of Android which the device is running.  This
update replaces the standard WebView implementation with Crosswalk, which means
that the dependency is a precise version of Chromium.  The version of Crosswalk
listed here as a dependency is a fork created by Medic Mobile which differs from
the standard implementation only in that the disk-usage quota restriction of 10%
of the available space, as imposed by WebView and XWalkView, has been removed.

Issue: medic/medic-webapp#2801